切换至mybatis

+++ src/main/java/com/cesium/demo/controller/CodeController.java
@@ -1,22 +1,15 @@
 package com.cesium.demo.controller;
 
-import com.cesium.demo.dao.CodeDao;
+import com.cesium.demo.mapper.CodeMapper;
 import com.cesium.demo.pojo.Code;
 import com.cesium.demo.utils.CopyFiles;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.core.io.Resource;
-import org.springframework.util.ResourceUtils;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.*;
 import java.net.URL;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Optional;
 
 /**
  * @author tanloo
@@ -26,15 +19,22 @@ import java.util.Optional;
 @RequestMapping("/code")
 public class CodeController {
 
+    private final CodeMapper codeMapper;
+
     @Autowired
-    private CodeDao codeDao;
+    public CodeController(CodeMapper codeMapper) {
+        this.codeMapper = codeMapper;
+    }
 
     @PostMapping
     public Object getCodesInfo(@RequestBody Long[] codes) {
-        Map map = new HashMap<Long, Code>(codes.length);
-        for (int i = 0, il = codes.length; i < il; i++) {
-            Optional<Code> code = this.codeDao.findById(codes[i]);
-            map.put(codes[i], code.orElse(new Code()));
+        Map<Long, Code> map = new HashMap<>(codes.length);
+        for (Long id : codes) {
+            Code code = this.codeMapper.getOne(id);
+            if (code != null) {
+                map.put(id, code);
+            }
+
         }
         for (Object code : map.values()) {
             String path = ((Code) code).getPath();
@@ -46,17 +46,30 @@ public class CodeController {
                         URL url = this.getClass().getResource("/static/gdal2srtmtiles.py");
                         String pyPath = new File(url.getPath()).getAbsolutePath();
                         url = this.getClass().getResource("/static/terrain_tiles");
-                        String tilePath = new File(url.getPath()).getAbsolutePath();
-                        System.out.println(pyPath);
-                        System.out.println(tilePath);
-                        File tileFile = new File(tilePath);
-                        if (tileFile.isDirectory()) {
-                            File[] fs = tileFile.listFiles();
-                            for (File singleFile : fs) {
-                                singleFile.delete();
+                        String tilePath = new File(url.getPath()).getAbsolutePath() + "/" + ((Code) code).getId();
+                        File tileFile = new File((tilePath));
+                        if (!tileFile.exists()) {
+                            boolean flag = tileFile.mkdir();
+                            if (!flag) {
+                                throw new Exception("创建文件夹失败，请检查权限");
+                            }
+                        } else {
+                            if (tileFile.isDirectory()) {
+                                File[] fs = tileFile.listFiles();
+                                if (fs != null) {
+                                    for (File singleFile : fs) {
+                                        boolean flag = singleFile.delete();
+                                        if (!flag) {
+                                            throw new Exception("创建文件夹失败，请检查权限");
+                                        }
+                                    }
+                                }
+
                             }
                         }
-                        //String cmd = "python" + pyPath + " " + pyPath + " --cesium --resume -z 0-10 -p geodetic F:\\Cesium\\terrain\\ASTGTM_N01E127_dem.tif " + tilePath;
+                        System.out.println(pyPath);
+                        System.out.println(tilePath);
+
                         String cmd = "cmd /c python " + pyPath + " " + pyPath + " --cesium --resume -z 0-10 -p geodetic F:\\Cesium\\terrain\\ASTGTM_N01E127_dem.tif " + tilePath;
 
                         System.out.println(cmd);
@@ -67,9 +80,8 @@ public class CodeController {
                         if (stat == 0) {
                             System.out.println("执行成功");
                             String s = "D:\\Administrator\\Desktop\\11ww\\terrain\\覆盖至生成tiles结果";
-                            String d = tilePath.toString();
                             try {
-                                CopyFiles.copyDir(s, d);
+                                CopyFiles.copyDir(s, tilePath);
                             } catch (IOException e) {
                                 e.printStackTrace();
                             }
@@ -77,12 +89,11 @@ public class CodeController {
                             System.out.println("执行失败");
                         }
                     } catch (Exception e) {
-                        System.out.println(e);
+                        System.out.println(e.toString());
                     }
 
                 }
             }
-            System.out.println(((Code) code).getPath());
         }
         return map;
     }
